CXP-1282: move a notifier in the new folder and fix test
