+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/Apps/handlers.yml
@@ -70,5 +70,5 @@ services:
         arguments:
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Security\ScopeMapperRegistry'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\SaveConnectedAppOutdatedScopesFlagQuery'
-            - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Service\AuthorizationRequestNotifier'
+            - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Notifier\AuthorizationRequestNotifier'
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\ScopeListComparator'
