+++ src/Akeneo/Connectivity/Connection/back/Infrastructure/Symfony/Resources/config/Apps/services.yml
@@ -47,7 +47,7 @@ services:
         arguments:
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Security\ScopeMapperRegistry'
 
-    Akeneo\Connectivity\Connection\Infrastructure\Apps\Service\AuthorizationRequestNotifier:
+    Akeneo\Connectivity\Connection\Infrastructure\Apps\Notifier\AuthorizationRequestNotifier:
         arguments:
             - '@Akeneo\Connectivity\Connection\Infrastructure\Apps\Persistence\FindAllUsernamesWithAclQuery'
             - '@pim_notification.notifier'
